Separated engine monitor and engine crawler (engine use 2 independant scripts now)
